Normalize SDL in a normalization step before validation
Currently, we perform transformations to normalize an individual
service's SDL during composition. One of the normalizations
(@extends directive) should actually be performed before validation
happens for simplicity.

The result of this realization was to pull out SDL normalization
into a separate step which happens before composition.
