+++ packages/apollo-federation/src/composition/composeAndValidate.ts
@@ -4,13 +4,15 @@ import {
   validateServicesBeforeComposition,
 } from './validate';
 import { ServiceDefinition } from './types';
+import { normalizeTypeDefs } from './normalize';
 
 export function composeAndValidate(serviceList: ServiceDefinition[]) {
+  const normalizedServiceList = normalizeTypeDefs(serviceList);
   // generate errors or warnings of the individual services
-  const errors = validateServicesBeforeComposition(serviceList);
+  const errors = validateServicesBeforeComposition(normalizedServiceList);
 
   // generate a schema and any errors or warnings
-  const compositionResult = composeServices(serviceList);
+  const compositionResult = composeServices(normalizedServiceList);
   errors.push(...compositionResult.errors);
 
   // validate the composed schema based on service information
